add search and report

+++ Backend/src/main/java/com/example/Backend/Repository/ReviewRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ReviewRepository extends JpaRepository<Review, Long> {
+public interface ReviewRepository extends JpaRepository<Review, Integer> {
     List<Review> findByProduct_ProductId(Long productId);
 }
